+++ packages/scoped-custom-element-registry/CHANGELOG.md
@@ -11,7 +11,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
-- registry in ShadowRootInit; matches current proposal but customElements
+- Added support for ShadowRoot.prototype.createElementNS()
+
+- Added the `registry` property to ShadowRootInit to match current proposal. `customElements`
   remains supported for compatibility
 
 ### Changed